Update entities to fix hibernate warning

+++ backend/api-service/src/main/java/at/technikum/boutique/hotel/apiservice/rooms/domain/RoomEntity.java
@@ -35,7 +35,7 @@ public class RoomEntity {
     private String description;
 
     @Column(name = "bed_amount", nullable = false)
-    private Byte bedAmount;
+    private Integer bedAmount;
 
     @Column(nullable = false)
     private Integer price;
